desktop-app/requirements.txt
pyinstaller==6.16.0
pywebview==5.4
